Extract AIs in parsers for bracketed AI data and DL URIs, not post-processing
Preparation for supporting unknown AIs.

In the parsers for bracketed AI element strings and DL URIs we know the
length of the AIs, as provided. Once these inputs are subsequently
converted to unbracketed data strings we can only differentiate the AI
from its value for known AIs in the AI table.